Move testMetadataOnlyQueries to separate class
This test cannot be reliably run in concurrent environment. We can
easily get a timeout as tests waits for resources to schedule beefy
query on shared query runner.
